Added app that builds a gpio library for the ArtyZ7 using the generic uio driver. This app is called 'libgpio'.
The app places the shared libraries in '/usr/lib' and places the the
specified include file in '/usr/include'.

This library can be used for gpios that are specified as generic uio
devices in the device tree. For more information about what
functionality is given by this library, please see the include file, or
'libgpio.c' in the recipe's directory.