Added app that builds a library for using the PWM ip-core from Digilent. This app creates the shared library and places the library files in '/usr/lib'. Additionally, the app places the include file for referencing the library in '/usr/include'. The library will be included in the rootFS by default.
The library supports PWM devices that are specified as generic uio
devices in the device tree. This project has a PWM driver connected to
the RGB LEDs on the Arty.

The library supports referencing the PWM ip-cores by their uio numbers.
Information about uio devices can be found in '/sys/class/uio'