Added app called 'libvrc' that builds a shared library for using kernel mode switching and framebuffers to control the output display on a connected monitor.
The app builds a shared library called 'libvrc' (VRC stands for Video
Resolution Control'. The shared library can be found in '/usr/lib' and
the include file is found in '/usr/include'. By default, this shared
library is included in the rootFS.

For more information on using libvrc, please see the recipe files or
check out the kmsdemo app.