Added recipe that builds 'gpioutil', a command line gpio utility.
This gpio utility can be used from the command line by using the command
'gpioutil'. The utility can be used to control gpio devices that are
listed as generic uio devices in the device tree. By default, this
project contains gpio drivers for the LEDS, BUTTONS, and SWITCHES on the
ArtyZ7.